% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{ar_parameters}
\alias{ar_parameters}
\title{Construct AR parameters}
\usage{
ar_parameters(
  coefficients = NULL,
  label = "AR parameter set",
  sign_convention = c("Deltares", "Standard"),
  order_tolerance = 1e-08,
  ...
)
}
\arguments{
\item{coefficients}{Coefficients from first to highest order.}

\item{label}{Label.}

\item{sign_convention}{Input convention.}

\item{order_tolerance}{Tolerance.}

\item{...}{Alternative coefficient input.}
}
\value{
An internal parameter object.
}
\description{
Construct AR parameters
}
