% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-classes.R
\name{ARForecast}
\alias{ARForecast}
\title{Autoregressive error forecast}
\usage{
ARForecast(
  parameters = ARParameterSet(),
  initial_errors = integer(0),
  series = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  time_step_minutes = integer(0)
)
}
\description{
Stores a projected AR error series and the inputs used to initialise it.
}
\section{Properties}{

\itemize{
\item \code{parameters}: Source \code{ARParameterSet}.
\item \code{initial_errors}: Input errors ordered newest first.
\item \code{series}: Forecast values as a \code{data.table}.
\item \code{time_step_minutes}: Duration represented by one model step.
}
}

