% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-classes.R
\name{ARMAResponse}
\alias{ARMAResponse}
\title{ARMA impulse response}
\usage{
ARMAResponse(
  parameters = ARParameterSet(),
  ma_parameters = integer(0),
  series = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })()
)
}
\description{
Stores a deterministic response to a unit residual impulse.
}
\section{Properties}{

\itemize{
\item \code{parameters}: AR component as an \code{ARParameterSet}.
\item \code{ma_parameters}: Active MA coefficients.
\item \code{series}: Response values as a \code{data.table}.
}
}

