% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{assess}
\alias{assess}
\title{Assess AR parameters}
\usage{
assess(x, ...)
}
\arguments{
\item{x}{Object to assess.}

\item{...}{Method arguments.}
}
\value{
An assessment result.
}
\description{
Assess AR parameters
}
