% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-tj-ar.R
\name{assess_jump_size}
\alias{assess_jump_size}
\title{Assess a proposed TJ-AR jump size}
\usage{
assess_jump_size(parameters, jump, noise_decorrelation_steps = NULL)
}
\arguments{
\item{parameters}{An AR parameter set.}

\item{jump}{Positive whole timestep jump.}

\item{noise_decorrelation_steps}{Optional estimate of noise decorrelation time.}
}
\value{
One-row \code{data.table} with jump-to-root ratios and warnings.
}
\description{
Compare a jump with characteristic-root timescales. This is a diagnostic, not
a universal pass/fail rule.
}
\examples{
assess_jump_size(default_ar_parameters(), jump = 5L, noise_decorrelation_steps = 3)
}
