% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-accessors-plots.R
\name{assessment_table}
\alias{assessment_table}
\title{Extract assessment table}
\usage{
assessment_table(x)
}
\arguments{
\item{x}{Assessment.}
}
\value{
Data table.
}
\description{
Extract assessment table
}
