% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{calculate_model_error}
\alias{calculate_model_error}
\title{Calculate observed-minus-simulated model errors}
\usage{
calculate_model_error(observed, simulated)
}
\arguments{
\item{observed}{Numeric observed values.}

\item{simulated}{Numeric simulated values of equal length.}
}
\value{
Numeric error values using \code{observed - simulated}.
}
\description{
Calculate observed-minus-simulated model errors
}
