% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-classes.R
\name{CharacteristicRoots}
\alias{CharacteristicRoots}
\title{Characteristic roots of an AR model}
\usage{
CharacteristicRoots(
  parameters = ARParameterSet(),
  values = complex(0),
  table = (function (.data) 
 {
    
    stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call. =
    FALSE)
 })(),
  time_step_minutes = integer(0)
)
}
\description{
Stores complex characteristic roots and their interpreted decay and
oscillation times.
}
\section{Properties}{

\itemize{
\item \code{parameters}: Source \code{ARParameterSet}.
\item \code{values}: Complex root values in polynomial order.
\item \code{table}: Root diagnostics as a \code{data.table}.
\item \code{time_step_minutes}: Duration represented by one model step.
}
}

