% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{decompose_ar}
\alias{decompose_ar}
\title{Decompose AR behaviour by root}
\usage{
decompose_ar(x, ..., initial_errors, steps = 480L, time_step_minutes = 15)
}
\arguments{
\item{x}{AR parameters.}

\item{...}{Method arguments.}

\item{initial_errors}{Errors ordered newest first.}

\item{steps}{Projected steps.}

\item{time_step_minutes}{Minutes per step.}
}
\value{
A long data table.
}
\description{
Decompose AR behaviour by root
}
