% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{detect_ma_order}
\alias{detect_ma_order}
\title{Detect effective MA order}
\usage{
detect_ma_order(coefficients, tolerance = 1e-08)
}
\arguments{
\item{coefficients}{Coefficients.}

\item{tolerance}{Activity tolerance.}
}
\value{
Integer order.
}
\description{
Detect effective MA order
}
