% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-et-ar-configuration.R
\name{event_ar_parameters}
\alias{event_ar_parameters}
\title{Construct ET-AR event parameters from catchment response time}
\usage{
event_ar_parameters(
  response_time_steps,
  middle_decay_steps = 5.203171,
  fast_decay_steps = 1/3,
  label = "ET-AR event parameters"
)
}
\arguments{
\item{response_time_steps}{Positive response time in model timesteps. For a
15-minute model, 48 steps represent 12 hours. \code{\link[=response_time_steps_from_tp]{response_time_steps_from_tp()}}
offers one temporary, approximate way to derive this from time-to-peak.}

\item{middle_decay_steps}{Decay time of the intermediate positive root.}

\item{fast_decay_steps}{Decay time of the rapid negative root.}

\item{label}{Description attached to the result.}
}
\value{
An AR(3) parameter object.
}
\description{
Retain the standard middle and fast roots while setting the principal decay
time from a catchment response estimate.
}
\examples{
event <- event_ar_parameters(response_time_steps = 48)
root_table(roots(event, time_step_minutes = 15))
}
