% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-lead-time.R
\name{lead_time_plot_data}
\alias{lead_time_plot_data}
\title{Prepare long-format lead-time plot data}
\usage{
lead_time_plot_data(x, common_period_only = TRUE)
}
\arguments{
\item{x}{An \code{ARLeadTimeResult}.}

\item{common_period_only}{Restrict data to the common valid period.}
}
\value{
Long-format \code{data.table}.
}
\description{
Prepare long-format lead-time plot data
}
