% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{plot_ar}
\alias{plot_ar}
\title{Plot a reach.postproc object}
\usage{
plot_ar(x, ...)
}
\arguments{
\item{x}{Object to plot.}

\item{...}{Additional arguments passed to a method.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a reach.postproc object
}
