% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{response}
\alias{response}
\title{Calculate an ARMA unit response}
\usage{
response(x, ..., ma_parameters = numeric(), steps = 480L)
}
\arguments{
\item{x}{Object defining the AR component.}

\item{...}{Additional arguments passed to a method.}

\item{ma_parameters}{Numeric MA coefficients ordered from \code{b_1} to \code{b_q}.}

\item{steps}{Number of response steps, including the impulse step.}
}
\value{
An \code{ARMAResponse} object.
}
\description{
Calculate an ARMA unit response
}
