% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arma.R
\name{response_components}
\alias{response_components}
\title{Compare response components}
\usage{
response_components(parameters, ma_parameters = numeric(), steps = 480L)
}
\arguments{
\item{parameters}{AR parameters.}

\item{ma_parameters}{MA parameters.}

\item{steps}{Steps.}
}
\value{
Long data table.
}
\description{
Compare response components
}
