% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{roots}
\alias{roots}
\title{Calculate characteristic roots}
\usage{
roots(x, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{...}{Method arguments.}
}
\value{
A characteristic-root result.
}
\description{
Calculate characteristic roots
}
