% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-accessors-plots.R
\name{aligned_series}
\alias{aligned_series}
\title{Extract aligned series}
\usage{
aligned_series(x)
}
\arguments{
\item{x}{Aligned result.}
}
\value{
Data table.
}
\description{
Extract aligned series
}
