% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-accessors-plots.R
\name{alignment_diagnostics}
\alias{alignment_diagnostics}
\title{Extract alignment diagnostics}
\usage{
alignment_diagnostics(x)
}
\arguments{
\item{x}{Aligned result.}
}
\value{
Data table.
}
\description{
Extract alignment diagnostics
}
