% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-forecast.R
\name{apply_ar_update}
\alias{apply_ar_update}
\title{Apply projected errors}
\usage{
apply_ar_update(
  simulated,
  ar_error,
  lower_limit = NULL,
  time = seq_along(simulated)
)
}
\arguments{
\item{simulated}{Simulations.}

\item{ar_error}{Error projections.}

\item{lower_limit}{Optional bound.}

\item{time}{Time values.}
}
\value{
Data table.
}
\description{
Apply projected errors
}
