% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{ar_parameters_from_timescales}
\alias{ar_parameters_from_timescales}
\title{Construct parameters from root timescales}
\usage{
ar_parameters_from_timescales(
  decay_times,
  oscillation_periods = rep(Inf, length(decay_times)),
  label = "Derived from timescales",
  tolerance = 1e-10
)
}
\arguments{
\item{decay_times}{Decay times.}

\item{oscillation_periods}{Periods.}

\item{label}{Label.}

\item{tolerance}{Tolerance.}
}
\value{
Parameter object.
}
\description{
Construct parameters from root timescales
}
