% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{assess}
\alias{assess}
\title{Assess AR parameter quality}
\usage{
assess(x, ...)
}
\arguments{
\item{x}{Object to assess.}

\item{...}{Additional arguments passed to a method.}
}
\value{
An \code{ARAssessment} object.
}
\description{
Assess AR parameter quality
}
