% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{calculate_model_error}
\alias{calculate_model_error}
\title{Calculate observed-minus-simulated error}
\usage{
calculate_model_error(observed, simulated)
}
\arguments{
\item{observed}{Observations.}

\item{simulated}{Simulations.}
}
\value{
Errors.
}
\description{
Calculate observed-minus-simulated error
}
