% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{convert_sign_convention}
\alias{convert_sign_convention}
\title{Convert coefficient sign convention}
\usage{
convert_sign_convention(
  coefficients,
  from = c("Deltares", "Standard"),
  to = c("Deltares", "Standard")
)
}
\arguments{
\item{coefficients}{Coefficients.}

\item{from}{Current convention.}

\item{to}{Required convention.}
}
\value{
Converted values.
}
\description{
Convert coefficient sign convention
}
