% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{default_ar_parameters}
\alias{default_ar_parameters}
\title{2024 EA default AR parameters}
\usage{
default_ar_parameters()
}
\value{
An internal parameter object.
}
\description{
2024 EA default AR parameters
}
