% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-roots-assessment.R
\name{effective_decay_time}
\alias{effective_decay_time}
\title{Effective decay time}
\usage{
effective_decay_time(decay_time, oscillation_period)
}
\arguments{
\item{decay_time}{Decay time.}

\item{oscillation_period}{Period.}
}
\value{
Effective decay.
}
\description{
Effective decay time
}
