% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{forecast_ar}
\alias{forecast_ar}
\title{Forecast an autoregressive error series}
\usage{
forecast_ar(x, ..., initial_errors, steps = 480L, time_step_minutes = 15)
}
\arguments{
\item{x}{Object defining the AR model.}

\item{...}{Additional arguments passed to a method.}

\item{initial_errors}{Numeric input errors ordered newest first.}

\item{steps}{Number of future model steps to calculate.}

\item{time_step_minutes}{Duration represented by one model step.}
}
\value{
An \code{ARForecast} object.
}
\description{
Forecast an autoregressive error series
}
