% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arma.R
\name{forecast_arma}
\alias{forecast_arma}
\title{Evaluate a deterministic ARMA recurrence}
\usage{
forecast_arma(
  parameters,
  initial_errors,
  residuals,
  ma_parameters = numeric(),
  previous_residuals = numeric()
)
}
\arguments{
\item{parameters}{AR parameters.}

\item{initial_errors}{Initial state.}

\item{residuals}{Innovations.}

\item{ma_parameters}{MA parameters.}

\item{previous_residuals}{MA state.}
}
\value{
Data table.
}
\description{
Evaluate a deterministic ARMA recurrence
}
