% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-accessors-plots.R
\name{forecast_series}
\alias{forecast_series}
\title{Extract forecast series}
\usage{
forecast_series(x)
}
\arguments{
\item{x}{Forecast.}
}
\value{
Data table.
}
\description{
Extract forecast series
}
