% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-accessors-plots.R
\name{lead_time_series}
\alias{lead_time_series}
\title{Extract lead-time series}
\usage{
lead_time_series(x)
}
\arguments{
\item{x}{Lead result.}
}
\value{
Data table.
}
\description{
Extract lead-time series
}
