% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-accessors-plots.R
\name{plot_lead_times}
\alias{plot_lead_times}
\title{Plot fixed lead-time results}
\usage{
plot_lead_times(x, thresholds = NULL, common_period_only = TRUE)
}
\arguments{
\item{x}{Lead result.}

\item{thresholds}{Optional thresholds.}

\item{common_period_only}{Common period flag.}
}
\value{
Plot.
}
\description{
Plot fixed lead-time results
}
