% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{response}
\alias{response}
\title{Calculate an ARMA unit response}
\usage{
response(x, ..., ma_parameters = numeric(), steps = 480L)
}
\arguments{
\item{x}{AR parameters.}

\item{...}{Method arguments.}

\item{ma_parameters}{MA parameters.}

\item{steps}{Response steps.}
}
\value{
An ARMA response.
}
\description{
Calculate an ARMA unit response
}
