% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{root_from_timescale}
\alias{root_from_timescale}
\title{Construct a modal root from timescales}
\usage{
root_from_timescale(decay_time, oscillation_period = Inf)
}
\arguments{
\item{decay_time}{Decay time.}

\item{oscillation_period}{Period.}
}
\value{
Complex root.
}
\description{
Construct a modal root from timescales
}
