% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{roots}
\alias{roots}
\title{Calculate characteristic roots}
\usage{
roots(x, ...)
}
\arguments{
\item{x}{Object for which characteristic roots are required.}

\item{...}{Additional arguments passed to a method.}
}
\value{
A \code{CharacteristicRoots} object.
}
\description{
Calculate characteristic roots
}
