% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-parameters.R
\name{roots_to_parameters}
\alias{roots_to_parameters}
\title{Convert modal roots to parameters}
\usage{
roots_to_parameters(
  root_values,
  label = "Derived from roots",
  tolerance = 1e-10
)
}
\arguments{
\item{root_values}{Modal roots.}

\item{label}{Label.}

\item{tolerance}{Tolerance.}
}
\value{
Parameter object.
}
\description{
Convert modal roots to parameters
}
