% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-alignment-lead.R
\name{score_lead_times}
\alias{score_lead_times}
\title{Score lead-time results}
\usage{
score_lead_times(x)
}
\arguments{
\item{x}{Lead-time result.}
}
\value{
Scores.
}
\description{
Score lead-time results
}
