% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-lead-time.R
\name{score_lead_times}
\alias{score_lead_times}
\title{Score fixed lead-time forecasts}
\usage{
score_lead_times(x)
}
\arguments{
\item{x}{An \code{ARLeadTimeResult} or compatible table.}
}
\value{
\code{data.table} containing MAE, RMSE, bias and improvements by lead.
}
\description{
Score fixed lead-time forecasts
}
