% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-tj-ar.R
\name{tj_ar_series}
\alias{tj_ar_series}
\title{Extract a TJ-AR projected series}
\usage{
tj_ar_series(x)
}
\arguments{
\item{x}{A result from \code{\link[=forecast_tj_ar]{forecast_tj_ar()}}.}
}
\value{
A copied \code{data.table} of projected error by lead time.
}
\description{
Extract a TJ-AR projected series
}
\examples{
# tj_ar_series(result)
}
